Rename pep zero generator files for consistency
