+++ pep_extensions/__init__.py
@@ -7,7 +7,7 @@ from pep_extensions.config import __version__
 from pep_extensions.pep_processor.html import pep_html_translator
 from pep_extensions.pep_processor.parsing import pep_parser
 from pep_extensions.pep_processor.parsing import pep_role
-from pep_extensions.pepzero.generate_pep_index import create_pep_zero
+from pep_extensions.pep_zero_generator.pep_index_generator import create_pep_zero
 
 
 # Monkeypatch StandaloneHTMLBuilder to not include JS libraries (underscore.js & jQuery)
